utils and double

+++ src/main/java/com/kishinskiy/Check.java
@@ -3,14 +3,13 @@ package com.kishinskiy;
 import java.io.File;
 
 class Check {
-
     /////// get used space on disk /////////
-        static int diskUsage(String diskName) {
+        static double diskUsage(String diskName) {
             File file = new File(diskName);
             long totalSpace = file.getTotalSpace() ;
             long usedSpace = file.getUsableSpace() ;
 
-            return (int) (((double)usedSpace/(double)totalSpace) * 100); // return percent of used disk space
+            return ((double)usedSpace/(double)totalSpace) * 100; // return percent of used disk space
         }
 }
